Do not wait for a grub screen after updating openSUSE installations
It is intentional and wanted that after the update process runs and yast triggers
a reboot, the newly setup system comes up immediately without passing through grub first.
This was wrong for a long time but got 'accidentally' fixed when yast2-bootloader was
corrected to no longer propose the bootloader configuration twice.