Updates: Allow to set YAST_SW_NO_SUMMARY as part of the job template
Until openSUSE 13.2, the default was to not show a summary after
installing packages. So it was correct to inject YAST_SW_NO_SUMMARY=1
into the upgraded systems. On 13.2, the default was changed and we can
no longer just inject it.

So the idea is to have the job template set it to 0 for the 13.2 update
case and keep the logic for the older cases.