The incredible fix for new LibreOffice
After tried many debugging way and all temporary fix all are failed in
ooffice test with new LibreOffice, unbelievable reordering the ooffice
test to after of oomath and oocalc test does work on openqa.o.o. This
sounds quite strange, but if this does work then just define it was
the workaround. The ordering of ooffice, oocalc and oomath is not that
important actually.