Fix 'ctrl-l being missed when too fast
This should fix some flaky tests depending on machine execution performance.

Replace all occurences of 'ctrl-l' to clear the screen in a console with
'clear\n' in form of the new helper function 'clear_console()' as 'ctrl-l' is
not buffered in the keybuffer whereas 'clear\n' is executed after the previous
command returns to the prompt.

Tested locally with "console-consoletest_setup" and following on x86_64. No
test of "ha" or similar has been done.

Alternative or complementary to
https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/923

Related issue:
https://progress.opensuse.org/issues/10138