products: sle: Distinguish versions of SLE
To support not only distributions in one branch but also versions the sle12
main.pm is moved to its own subfolder. As "os-autoinst" does not know the
notion of a "VERSION" regarding test cases we need a little hack to load the
tests from another "main.pm" and also rewrite the "DISTRI" variable as it is
used to discover the needles directory within os-autoinst after loading the
test definitions from main.pm.

This will change the value of "DISTRI" within vars.json in each run.