zdup: 13.2 to TW: Network changes naming scheme
in openSUSE 13.2, systemd was configured to use predictable names.
This did not work with the virtio-net driver used inside openQA.

Once the update to TW is complete, predictable names start working (systemd fix)
but results in this constellation with a no-longer configured network

Catch that case, copy eth0's config to be ens4's before starting zypper dup.