13.2 Update case: copy eth0 config to ens4
openSUSE 13.2 was configured for predictable names but this did not work
for virtio-net devices. A maintenance update could 'fix' this, but it would
break existing systems.

In Tumbleweed (systemd >= 228) the virtio-net driver is now matching the
system configuration and the device is thus renamed from eth0 to ens4.

Re-apply the configuration as it was used prior the update and restart the network.