yast2_i: use zypper -n -i (ignore unknown packages)
The purpose of the zypper call is to ensure that all packages listed are
removed from the system. In some cases, packages are not installed and
not available in the limited set of repositories, resulting in zypper
abortnig the task instead of removing the remaining packages.

Using zypper -i ignores the unknown packages and performs the task on
the remaining set.

This only makes sense for openQA on package removal - installations we have
to take care of in a better way