Force 1024x768 after reboot
As explained in commit d75924cf3d18eafb5ff8be372064654ea32b9e82 we need
to make grub come up in 1024x768 to get the proper resolution when kms
is not used.

We also need to reset the EFI boot partition to work around how KIWI
does things and a bug in shim-install.

https://bugzilla.suse.com/show_bug.cgi?id=968270
https://bugzilla.suse.com/show_bug.cgi?id=968264