Update evolution_mail_ews.pm to fix address book auth failure
Sometimes the evolution_mail_ews.pm test case fails at step of
fetching the Global Address Book list, because it indicates
that the authentication is required but missing in test script.
Strange is the auth isn't really required for fetching the list
of address book when evolution is run in another environment.