Add entry for FIPS test suites in main.pm
The "FIPS setup" test suite is the basic one to make the image
for FIPS tests afterwards. The setup test suite must be run with
FIPS=1 variable to make sure the FIPS pattern is enabled during
image installation. And other FIPS test suites should be run upon
the generated image.