Use more explicit "boot_to_desktop" as last step of post_zdup
The migration and upgrade tests are expected to reach a fully working desktop
session which was handled by "wait_boot". As we recently failed tests a lot
in this state even though the issue had nothing to do with the migration or
upgrade itself this should be made more explicit by using the test step called
"boot_to_desktop" to resemble this.