Extend installation timeout gracefully on aarch64
aarch64 installations often timeout. The reason so far is unknown, bugs have
been filed:
https://bugzilla.suse.com/show_bug.cgi?id=982136
https://bugzilla.suse.com/show_bug.cgi?id=982137

This change catches the case when on aarch64 the initial installation timeout
of 2000 s is hit, records a soft failure with a reference to the
aforementioned (first) bug and continues for up to another 2000 s. If the
installation is still not complete, the test will then hard fail.  No change
to the timeout for other architectures has been done.

Verification test run: http://lord.arch/tests/238