Add helper function assert_screen_with_soft_timeout
Change all occurences of assert_screen within console and x11 tests with a
timeout of less than the default (30s) to use a "soft timeout" first with the
old timeout value and then, if the soft timeout is hit, continue up to the
full timeout. See description of assert_screen_with_soft_timeout for details.

This should help in occassions where a module failure is more often than not
regarded as a "flaky issue" and not cared about but the job is simply cloned
to retry.

Related issue: https://progress.opensuse.org/issues/12110