products: opensuse: Run extra tests even on server-like scenarios
Most of the extra tests do not actually need a desktop environment so
we modify the logic in that subroutine to allow them to be executed
even on server-like jobs. For any test that really needs a DE a new
subroutine was added (any_desktop_is_applicable()) which can be
used to check if the DESKTOP variable is set properly.

Signed-off-by: Markos Chandras <mchandras@suse.de>