bootloader_s390: Always show debug regardless of result
This allows easier comparison of "good" and "bad" state. Also changing
"script_run" within show_debug() to type_string to not rely on a potentially
broken serial port which we want to debug.

The assumption about return code of script_run was wrong. Of course it does
return the exit code of the executed script, not a perl true/false value.