susedistribution: Fix loading within consoletests from qemu snapshots
Add back what was working before we reworked the console handling.

There should not be a check_screen with no alternative so I am removing the
check for a second return to be pressed in case a user is not logged in.
This is just a waste of time and should not be necessary at all.

Now checking for already logged in shell, i.e. starting with the single '$' as
prompt. Needs updated needles.

Local verification done in the two cases of loading from snapshot when the
session is already logged in as well as when booting a clean machine where the
login is still necessary.