Fix yast2_snapper for SLE 12-SP1
Last change to yast2_snapper triggered failures which are actually correct as
the deleted screenshot is still visible if looking "fast enough". The old
version did assert_screen on the main window first and then look if the "to
be deleted" screenshot is really gone, which it normally was at this time but
not directly after returning from the delete-dialog. The newer test is faster
and actually finds the snapshot still shown before it vanishes after the next
action. This change applies a workaround for the released product version
where this problem has been observed. If necessary this can also be extended
for other products and corresponding versions.

Also speeding up the look through the list by going to the end of the list
first and then go up instead of stepping through the whole list from top.

Verification run: http://lord.arch/tests/2097