toolchain: Configuring and compiling can take way longer
Should fix issues as in
https://openqa.suse.de/tests/487089

In general configuring and compiling time scales a lot with performance of the
underlying platform. The test is mainly not about performance and openQA is
not well suited to test this so the timeout values should be way less
restrictive and only a last resort. For the sake of simplicity the same robust
value is used independant of architecture or platform as we don't really have
reliable performance figures which are only architecture but not worker
specific.

Related progress issue: https://progress.opensuse.org/issues/12896