Create new test module encrypted volume activation
When the variable ENCRYPTED_CANCEL_EXISTING is set, it will cancel the
activate encrypted volume prompt which appears during installation to a
storage device with existing encrypted lvm volume. A workaround is
implemented for bsc#989770 which causes the activation prompt to be
displayed twice.

When the variable ENCRYPTED_ACTIVATE_EXISTING is set it will enter the
password for the existing volume to activate it.