yast2_snapper: Make die a soft failure with known cause
It looks like it can always be the case also in more recent products that the
snapshots requested for deletion stay around for some time even though the
dialog already disappeared. As we have a bug reference now we can track this
but try more graceously to wait for the snapshot being deleted. This will also
help work on the bug as obviously the question is if the snapshot will
eventually be deleted at all.