zypper_lifecycle: Do not rely on date as getting date is non-atomic
The  being evaluated on the worker and not the SUT is suspect to timezone
differences, e.g. see
https://openqa.suse.de/tests/506437#step/zypper_lifecycle/29, where this
fails. But even when we would get the time from the SUT and then compare it
against the zypper lifecycle output we could hit the day boundary. It is
better to not try to parse the date at all.