cryptlvm: Introduce ENCRYPT_FORCE_RECOMPUTE
Introduces a variable to force an encrypted partition proposal to be
generated regardless of whether the default partition proposal is an
encrypted one. If the default partition proposal is not encrypted it
will now signal a soft failure whether the variable is set or not.

Also ensures a screen shot of the radio buttons in the partitioning
proposal is taken.