Get more logs from fortran test in case of failure
https://openqa.suse.de/tests/515119#step/gcc5_fortran_compilation/24
produced error but from the information there I am unable to determine
what might be the problem. Be it test, openQA, or system. This commit
removes some tests which we are not able to verify automatically, makes
verification of the rest of the tests saner and uploads more logs to
openQA.