Update LLVM and LTP, do not let fails unnoticed
Update of LLVM/Clang and LTP test suites. Adapt to out-of-source-tree
build. Run LLVM to Clang test suite.

Previously test was run with `script_run` so fails got unnoticed. That's
changed with `assert_script_run`. That also means that if some test
cases in e.g. LTP fail, the test fails as well (LLVM/Clang seems to pass
all the time atm). So, we need to either fix the tests or do not run
them if they are deemed unstable:
http://assam.suse.cz/tests/2769#step/gcc5_C_compilation/19 (build and
run logs attached in Assets).