Add compatibility of smt registration
Use 'SMT_URL' variable rather than 'SCC_URL' for
smt registration, because of the behavior is not
very identical between scc and smt registration.

No need to input registration codes for smt registration.

And the behavior of smt registration on 12sp1 is
a little different, we have to add a needle to check it
and then manually send next to start registration after
importing smt server's certification.