For kernel updates we need to ignore -base packages if in conflict (#1786)
For this to work, we need to make the whole solver test case a bit more complex
as it's possible to install kernel-default-base-5 and kernel-default-4 but not
version 5 of both. But this is what we need to do to make the install_packages
test useful. So install the latest solvable instead of the selectable and remember
what we asked for so we can backtrack later

(and perltidy the script as we changed more of it)