Fix SLE modules do not have release notes
While installing the Web and Scripting Module (WSM), openQA fails with
'NO matching needles for release-notes-wsm'. However, this should be
skipped because there are not release notes for WSM. The same applies
to all the modules.

Also, I removed the bugzilla link because this is a comment in a
RESOLVED bug which does not add a lot of information. From what I
understood there is no release note for WE.