Add a metadata boilerplate to every test
This is extracted from git and often enough nonsense - or duplicating
infos already present in the file. But if we want to enforce a policy
on having this information in the file, we need to make it easy for
contributors to see how it's supposed to look like.

So I add G-Summary and G-Maintainer, that the next contributor to the
test needs to adapt to be Summary if correct - or change the text while
at it.