Prevent 'sshboy' messing up unrelated tests
The user sshboy is created by the sshd test module and used only in sshd
itself and ssh_pubkey (FIPS) for now. To prevent confusion of reviewers and
unrelated tests (e.g. log into displaymanager) the user is removed in its own
test module regardless if the sshd test module failed or passed in before.