Always reboot to ensure a properly working updated system (bsc#989696)
In case of online migration in bsc#989696 we hit the case that the migrated
system does not boot. The old SLE 12 GA kernel version was suspected to be the
case. `zypper patch` would pull in an updated kernel but it is never applied
as there is no reboot between the `zypper patch` call and the whole step of
online migration.