Update aide_check.pm: Increase timeout
The default timeout, 30s, is sometimes not enough to
install the aide package, extend it to 90s. Also set
timeout of "aide -i" and "aide --check" commands to
be 60s to avoid test failure caused by timeout.

Replace assert_screen with validate_script_output to
check the output of "aide --check" command, which will
improve the stability of the test case.