Better place for ImageMagick test in main.pm
At the moment, the ImageMagick test is never triggered. Since eog is
used in the ImageMagick test, it can be triggered right after the
'eog' test. Normally, the SDK needs to be installed but at this stage
the host is either desktop or with WE and both provide ImageMagick
along with SDK.

Now, the ImageMagick test will be included in the qam-gnome testsuite
in openqa.suse.de.