LAMP(php) test (#1966)
Install LAMP and test the basic functionality of PHP
(apache2-mod_php5), the connector between PHP and MySQL (php5-mysql),
the connector between PHP and PostgreSQL (php5-pgsql), and
shibboleth-sp.

This commit enables also the conosole/postgresql94server test in
main.pm

Those test are all triggered when DESKTOP is set as 'textmode'. The
php5* tests are triggered when the wsm module is also defined either
in ADDONS or in SCC_ADDONS variables.