Keep the order of the actions taken in firefox test
Originally we had the steps taken to start firefox hard coded
into check_screens that timed out in doubt. I changed it to run
assert_screen on multiple tags, but that had the problem that
we missed the default_browser because the match_has_tag returned
the reader-window needle.

Now fixing this by not using match_has_tag but an check_screen without
timeout to check the screen found by assert_screen for an explicit
tag.