Get rid of REPO_SLES_DEBUG, visual glitches fixes (#1999)
* Better handle screen distortion from previous test

On PPC64LE, at least, the test which runs before `crash`, creates
post-mortem screen distortion. Previously we tried to clear the screen
before needle matching, but a more robust solution is to restart the
system and then do the matching. To achieve that code was reshuffled
here and there.

* Construct FTP dir name from REPO_0 variable

REPO_SLES_DEBUG variable formerly provided by `rsync.pl` script is now
not available. The information required - a directory on FTP server
which holds kernel debuginfos - is now get from REPO_0 variable and
couple of string transformations on it's content.