Bootloader/Text mode selection: Press 'ESC' on the new video mode menu
When selecting text-mode, the new menu no longer closes when pressing
'ret'. This is done as there is also a 'color scheme' submenu which a
user might want to pick right away too. Instead, it is now needed to
press 'ESC' to close the menu and get back to the grub menu.

We can't unconditionally press 'ESC' as in case the menu would close
directly with 'ret' (old style), the keypress would prompt grub to ask
to quit to the grub console.