Fix login_test occasional failure
The login_test types 'settings' in the overview of gnome and selects
it with assert_and_click. Although a delay is already set in
typing, it is not adequate. Recent test failed because the assert
happened before the full string was typed. So a 'wait_still_screen'
after typing fixes this issue.