Add direct/native LTP and virtio_console tests
This includes two tests, one is the LTP native or direct test
runner. The other is a simple test case for logging into a virtio serial
terminal. Both require the virtio serial terminal patch to os-autoinst,
although the LTP test cases can be ran on a VNC based console, the patch
is still required for the is_serial_terminal subroutine.

Currently the LTP is ran through the CTCS2 test harness in OpenQA and a
shell script is provided upstream for manual or automated testing. These
test cases remove those dependencies by parsing the LTP runtest files in
OpenQA and executing the test cases by individually 'typing' them in on
a terminal.