Expect installer self-update to be disabled by default
Using the new variable "INSTALLER_SELF_UPDATE" the installer self-update
feature can be explicitly enabled.

Actually testing if the installer self-update is not easy and so far is only
done implicitly by assuming the window over an ssh+X installation to be not
fullscreen which we have covered so far only on s390x zVM installations. This
check is now also changed to check for the explicit enablement of installer
self-update, disabled self-update as well as providing a fallback when the
window is not fullscreen in the default case which can mean that the
self-update was set as default again which should not happen unnoticed.