Ensure sle11 is self-contained
This is very much desired to know which version was tested of all the test
plan and the modules. Also, it means for old SLE11 we do not need to care
about support in the modules anymore and can delete modules that are not used
within the master branch which would otherwise be way harder to check for.

Done by branching of sle11 properly from master and additionally include
"main.pm" from sle11 branch, commit 0c73f14.