Refactoring and dedup evolution cases
coolo has submitted three commits to dedup the evolution cases in PR#2100
 - commit 3174652
 - commit 8af1cb5
 - commit 5614eee
But this PR doesn't solve all the problems since he is lack of time,
I took over it and polished the previous three commits to make it work.

So this is the fourth (hope it's the last) commit to dedup evolution cases.

  see also: poo#13118