Test seahorse only as part of EXTRATESTS
Seahorse, as useful as it is, has a very questionable status at the moment
and upstream is fully aware of it. It randomly fails to grab focus in the
right moment - and having it in the default test group only causes grief
and random failures at no benefit.

As we don't want to lose the test completely, we move it to the EXTRATESTS

https://progress.opensuse.org/issues/15064