Change kernel-version regex to be more permissive (#2223)
The kernel version got longer by one block which causes the sed statement
to not match. This commit makes the regex more permissive.

Attention: This will not work for kernel-rt, because it has "rt" in the
version number but since there are no live patches for this flavour it is
only a minor concern.