Improve empathy_irc test
The empathy_irc test expects that shortcuts fail on SLE12-SP2 and succeed on other version. However, the test should record a soft failure when the test actually fails and not based on the version. Also the workarounds are now used in previous versions as well, since the shortcuts might fail occasionally (openQA lag).

+bugfix: new needle - assert before typing