Do not re-create VM domain after installation
POO#15784

The aproach with installing SUT, destroying it, and re-creating it with
the installed image does not produce reliable results, frequent VNC
stalls.

On KVM it's just fine to wait after installation for the new system to
boot, on Xen - as the VNC connection is closed after restart by host -
we have to find the right moment when the SUT is booting.

Verification run on KVM, Xen HVM, and Xen PV respectivly:
http://assam.suse.cz/tests/4616
http://assam.suse.cz/tests/4626
http://assam.suse.cz/tests/4627