LTP: networking: temporary define RHOST to bypass mrshd security check
This fixes various scripts from net.tcp_cmds which use rsh.

We need to define RHOST as mrsh daemon for security reason requires for
this connection call hostname as '127.0.0.1' or 'localhost' (e.g. not
hostname).

NOTE: This is an exception as tests which use test_net.sh cannot have
RHOST defined as they use network namespace.