Workaround for boo#1019090, bnc#1014602
boo#1019090: if there are 'java-bootstrap' packages,
try to install them, unless they conflict with java.

bnc#1014602: java 1.7.0 is reported by java when having ibm-java 1.7.1
installed. This behavior is expected and acceptable according to IBM.

Also only fail the test for 'javac' if the -devel package should be
installed (SDK available) but it's not.

Update the date in the comment section (2016 -> 2017).

Replace all the undefined exit codes with 1.