Verify USB repo is disabled
Initially bsc#1012258 was raised to fix the issue as seen in
https://openqa.suse.de/tests/643667#step/zypper_lr/12
that the USB repo is disabled as it should which was not foreseen by the test.
This issue has not been causing test failures because of a change how the
device is called by linuxrc. There is also another bug which has been raised
and will be tracked for the time being as soft failure: boo#1019634