INSTALLONLY should not apply updates
Applying updates and also testing this is a good idea but the "install only"
jobs are not the correct location.

An update notification like in
https://openqa.opensuse.org/tests/332430#step/consoletest_finish/11
could be better fixed by needles which either explicitly accept the update
notification or mask it so that it does not matter because we should not care
at the state of "install only" if there are any updates or not.

Any job using INSTALLYONLY should test only if the SUT can boot correctly.
That is especially true for tests stating BOOT_TO_SNAPSHOT which boots into a
snapshot that is read-only. Trying to apply updates does not make sense after
booting into a read-only snapshot.

This reverts commit 40a89e4be9f0367a3c051d869d6586cfeacae73a.

Verification runs:

* kde @ TW: http://lord.arch/tests/5592
* gnome-image @ Leap 42.3: http://lord.arch/tests/5593
* gnome-gdm @ Leap 42.3: http://lord.arch/tests/5594
* gnome maintenance @ Leap 42.2: http://lord.arch/tests/5595

Related progress issue: https://progress.opensuse.org/issues/16018