Do not waste time checking for untrusted key
This change will prevent needless waiting for 10 seconds in the normal case
when no untrusted key pops up to be be trusted. In light of bsc#990254 we want
to measure the time for downloaded release notes to pop up. Waiting an
additional 10 seconds is not really helpful in that regard.