LTP: Prevent race condition by sending ret after read
If os-autoinst pauses between type_string and wait_serial, but the SUT is able to
run, then the SUT's output will be consumed by the wrong call to
wait_serial. We avoid this by sending return after consuming the command echo
so that the command will only be run just before, or during the correct call
to wait_serial.