Do not run all consoletests for Krypton and Argon
Those Live medias are of interest to the UI users of the respective
desktop. Testing the entire suit of console tests only gives a lot
of fails/softfails, which the Desktop Envrionment Teams do generally
not care for.

Skip the generic consoletests with the exception of the very basic ones.