Sleep as workaround before typing in editor/xterm (poo#17442)
I really do not like the sleep and want to get rid of it again but I do not
see a better choice as a workaround. `wait_still_screen` does not work because
of blinking cursor, `wait_screen_change` does also not apply because we have
an `assert_screen` just in before. `wait_idle` should be avoided, no choices
left.