Disable testing 'java-1_9_0-openjdk' in TW
The 'java-1_9_0-openjdk' is not considered ready for public release
so there are bugs that the developers are aware of and they are
currently working on them. As a result there is no point patching
workarounds against a not-ready-to-be-released version.

The test for 'java-1_9_0-openjdk' will be enabled again at
'27.7.2017' -- General Availability for Java 9.