Prevent typing too early into shell
This was not really causing an error but confusing screenshot. Typing before a
bash prompt becomes ready and executing that afterwards is valid behaviour but
we can wait a little bit more by waiting for the prompt to be ready which is
also what most users would do anyway.

Fixes https://openqa.suse.de/tests/832883#step/transactional_update/26