LTP: Try to install GCC 32-bit binaries
Soon LTP will support -m32 (32-bit compat mode) for individual tests so we
need to install the 32-bit library binaries and headers for glibc and
GCC. Instead of listing all the necessary libs, we can just install
gcc-32bit. If gcc-32bit is missing we do not want to fail here, the LTP build
system will deal with it and we will get TCONF for the effected tests.