Wait for screen change after pressing key
https://openqa.suse.de/tests/878705#step/partitioning_warnings/14
I guess this failed because loop was so fast that it catched same screen twice.
In log needles matched with same tag but video looks like the second match was
partition-warning-no-swap which ended loop despite log is showing match
partition-warning-too-small-for-snapshots twice.
Perl is so fast or qemu so slow or both simultaneously that even logs *LIE*!