LTP: Yet another fix for pidstat
While Bash is outputting background job information (after using '&') any new
input is discarded. This results in half a git clone command typed in.

Apparently wrapping the command in brackets prevents any input/output to
STDIN/OUT with '&', but also means we have to user 'ps' to get the PID instead
of %1.