wait_boot: Extract wait_for_boot_menu helper
Extract helper method to reduce complexity of wait_boot and as preparation to
make more code reusable from code which is currently duplicating reboot/boot
content.

Changes behaviour for OFW slightly by immediately selecting the selecting boot
entry of the grub menu instead of waiting for the timeout to boot it.