LTP: Wait for the terminal prompt to appear before typing
I suspect Bash drops I/O data on rare occasions because we are sending input
too early. If true then we need to wait for the prompt to appear before
sending data.

Also we may accidentally send data to the STDIN of the echo program if we are
very quick. I don't think echo reads STDIN, but it's possible that the data
will get dropped when echo closes, depending on how the shell handles STDIO
for its children.

If this stops some of the random failures then we may need to do it for all
script_run calls. This makes the output from the LTP look even worse, but it
is more important that the tests are run.

Requires os-autoinst PR: https://github.com/os-autoinst/os-autoinst/pull/772