Allow to trigger 'install_all_from_repository' by defining variable
We have the test module 'install_all_from_repository' which has been added to
the test plan explicitly for HPC testing. For generic testing it would be
useful to enable the test module without needing to change test source code.
This commit triggers the test module at the end of console tests if the
variable INSTALL_ALL_REPO is defined in the test. For a successful execution
the variable must contain a valid repository name as described in the method
description for "install_all_from_repo".