Simplify autoyast installation logic
The long error handling parts within the "happy path" confused many reviewers
of test results. Using a post_fail_hook that can be separated in a cleaner
manner. Also, replacing the 'while(1)' with a 'do...until' can reduce
complexity.

The logical flow changed about the handling of the timeout values.in a way
that there is either a hard-fail in case of unexpected screens. Also, in case
a dialog is encountered the timeout value is not changed but the corresponding
assert_screen is restarted with the original timeout. That means in case of
additional dialogues potential regressions in overall execution time are not
catched in favor of having simpler test code.

Verification run: http://lord.arch/tests/6238