Skip firefox_audio test on Rescue-CDs
Space is getting more and more of an issue on the rescue CDs - and providing
multimedia capabilities is not high on the list for the Rescue disk.

We rather sacrifice a few MB of sound drivers from the kernel to get usable
recovery tools on the media - as a consequence, we can't successfully test
audio in firefox though.