Fix poo#19234: Add boot=/dev/xyz for fips_lvm_encrypt
According to bsc#989491, the encrypted LVM fails to boot in fips
mode because in the initial ramdisk the /boot is not populated or
mounted. Adding the "boot=/dev/somewhere" option to the kernel
line is required. It must point to the device containing "/boot",
so that the initial ramdisk can find the HMACs.