Prevent kontact comming up again after reboot
From discussions with the KDE team the behaviour which was previously random
but now in current staging:C very consistently kontact can persist in the
background as a process. After re-login/reboot the application would be
started which is not expected in our tests. This behaviour seems to be
consistent now with the move to gcc7.

For more discussion around this also see
https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/2962

Verification run: http://lord.arch/tests/6331