Fix still busy filesystem on wipefs in some occassions
Not using a lazy unmount could work in some cases but in most it would still
fail because the filesystem is busy from the previous filesystem experiments.
So we instead force the wiping of the filesystem. Also later in other test
modules when the disk variable is still there we reuse it. It must be the same
disk anyway. This way if the disk looks still like used or not unpartitioned
it's forcefully overwritten as a last resort.

This is a followup to
https://github.com/os-autoinst/os-autoinst-distri-opensuse/pull/2964

Verification run: http://lord.arch/tests/6418

Related progress issue: https://progress.opensuse.org/issues/19370