Improve comment why have console=tty as boot parameter
See poo#19398 and bsc#1011815. The problem is that sometimes kernel
messages are printed to active console and prevent needles from
matching. We already redirect kernel messages to serial device and do
not need them on tty, whereas installation cannot start without tty
defined as system console, hence we only can workaround this by either:
1) reducing loglevel
2) using klogconsole
3) changing grub config on installed system

As it's only nice to have, we leave it as it is, and change the tests
only if it becomes critical issue.