Test Firefox's java support in opensuse
In one of the releases java was not supported in firefox. We already
have such regression test for sle, which was migrated to TW and Leap.
This commit includes needles required to the test for TW and Leap 42.3
on gnome.

Progress reference: https://progress.opensuse.org/issues/18850
Verification run Leap 42.3: http://gershwin.arch.suse.de/tests/102
Verification run TW: http://gershwin.arch.suse.de/tests/101

Move firefox tests from version specific folder

As part of poo#18850, we move tests from sle12 which is version
specific, as we migrated one of the tests for TW and Leap and may do so
for other versions, as well as for other versions of sle.