Fix zkvm login to consoles in consoletest_setup introduced by 93d024a9
After the recent change in os-autoinst the login handling was changed but went
a bit too far as it seems. On zVM we need explicit password handling. On zkvm
we are already logged in in this case but the shell might not show up
immediately after the console switch is done. The assert_screen looking for
multiple alternatives already covers this correctly.

Verification runs:
* zkvm gnome: http://lord.arch/tests/6504#step/consoletest_setup/33
* zkvm textmode: http://lord.arch/tests/6506
* zkvm toolchain_zypper (not using consoletest_setup): http://lord.arch/tests/6505

Related progress issue: https://progress.opensuse.org/issues/19486