Remove redundant yast2_lan_restart triggering in yast2_ui
yast2_lan_restart was triggered twice in extra_tests_in_gnome as well as
yast2_ui where it is the only module consistently failing. Also it seems to be
more the network restaring behaviour, i.e. with wickedd, and not solely about
the yast UI so it is better to call it as parts of extra tests.

Related progress issue: https://progress.opensuse.org/issues/19296