Add toolchain lyfecycle test
See fate#322050 and poo#15306. Detailed test case can be found in fate
ticket. We already test zypper lifecycle command, but it's required to
extend existing tests and perform tests which are specific for the
module. Here we implement test for toolchain module packages. Test is
executed in the test suite where toolchain module is already installed.