Add wait for snaps list in user_defined_snapshot
In user_defined_snapshots tests we press down key until match needle
with required snapshot, whereas on slow system we start pressing down
key before the list is even shown. It results in unexpected initial
position of cursor in the list and we go only one direction.
In this commit we verify that screen was changed and that snapshots list
is currently shown. Needles for aarch/x86 and PPC are added.