Soft-fail when gdm cannot start from snapshot
See poo#19762 and bsc#980337. On ppc and arch during
user_defined_snapshot execution, when we boot from snapshot system
stucks in one screen, whereas if we switch to tty2, test can continue.
As it's not clear how long will it take to fix the issue, we introduce
soft fail here.