Fix scrolling to the end of snapshots list
Before user_defined_snapshot test we execute yast2_lan_restart test case
which triggers creation of many snapshots, that the list doesn't fit
without scroll. In previous implementation we were pressing pgdn button,
which doesn't scroll to the end of list to verify that test snapshot is
not there. It's questionable if we need this check at all, but with
"end" button we will navigate to the end of list.