Introduce timeout for installation
As part of commit 44e5ffe3 we have removed timeout in order to get
better error handling if test fails to match any of possible needles.
Whereas under some conditions, we may end up in infinite loop if
assert_screen always matches a needle. So, we introduce timeout back,
but try to keep advantages of previous implementation as well.