Disk 'b' always the secondary disk for tests
This change does two things: (1) Disk 'b' is always content of bash
variable `disk`. (2) The disk is not required to be empty; it's up to
the test to have it in expected condition.

Ad (1) I don't see in which scenario it would be disk 'a'.
Ad (2) Seeking unpartitioned disk (of 'unknown' partition table)
requites previous tests to clean after themselves; snapshots reseting
environmental variables as enviromnent changes (e.g. `systemctl rescue`)
come into play.

Thus also renaming the function as it changed purpose.