Rewrite `snapper --no-dbus` tests for Tumbleweed
poo#19804

In Tumbleweed we can't stop DBus on demand and even if we can do it in
SLES and Leap, it's quite dubious to have system without DBus. Safe way
to test `snapper` w/o DBus support (`--no-dbus` option) is to switch
boot target from default to rescue and test from there.

Unfortunatelly in rescue.target we don't have networking, which depends
on DBus (in Tumbleweed anyway), and thus we are unable to use
`script_output` and rewrite by using environmental variables is not
entirely straightforward. So, as a result some tests are now not being
run with `--no-dbus` option, but the whole set is not SLE15-ready.

Two support functions were added to setup (`snapper_nodbus_setup`) and
restore (`snapper_nodbus_restore`) from rescue.target environment.

`console/snapper_thin_lvm` is now enabled on Leap & Tumbleweed.